Add autodetection for calendars
When you provide the server-hostname and user-credentials, the plugin is
now able to autodetect the calendars. This works with the plain
server-url but also for reconfiguring with a complete path to a specific
calendar.
Not the most fancy ui but it works, at least for the test with a
productive owncloud server